requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
psycopg2-binary==2.9.5
requests==2.25.1
sqlalchemy==1.4.41
python-dotenv==0.19.2
typing_extensions==4.6.3